+++ packages/lib-sourcify/src/lib/verification.ts
@@ -20,12 +20,7 @@ import {
   decode as bytecodeDecode,
   splitAuxdata,
 } from '@ethereum-sourcify/bytecode-utils';
-import {
-  getAddress,
-  getCreateAddress,
-  keccak256,
-  id as keccak256Str,
-} from 'ethers';
+import { getAddress, keccak256, id as keccak256Str } from 'ethers';
 import { hexZeroPad, isHexString } from '@ethersproject/bytes';
 import { BigNumber } from '@ethersproject/bignumber';
 import { defaultAbiCoder as abiCoder, ParamType } from '@ethersproject/abi';
@@ -688,16 +683,6 @@ export async function matchWithCreationTx(
         abiEncodedConstructorArguments;
     }
 
-    // we need to check if this contract creation tx actually yields the same contract address https://github.com/ethereum/sourcify/issues/887
-    const createdContractAddress = getCreateAddress({
-      from: creatorTx.from,
-      nonce: creatorTx.nonce,
-    });
-    if (createdContractAddress.toLowerCase() !== address.toLowerCase()) {
-      match.creationMatch = null;
-      match.message = `The address being verified ${address} doesn't match the expected ddress of the contract ${createdContractAddress} that will be created by the transaction ${creatorTxHash}.`;
-      return;
-    }
     match.libraryMap = libraryMap;
 
     match.abiEncodedConstructorArguments = abiEncodedConstructorArguments;